==> ETL/1_Url_ResT_WebScrapping_Etapa1.ipynb <==
Notebook para extraer las Url de los restaurantes de Madrid

import requests
import pandas as pd
import os

import pickle
import helium
from time import sleep
from bs4 import BeautifulSoup 


#Con esta pagina, es el punto de inicio para escoger los restaurantes 20 destinos mas populares de España 
url = "https://www.tripadvisor.es/Restaurants-g187427-Spain.html"
browser = helium.start_chrome(url, headless = True)

# Hacemos .page_source para tomer el html de la página
soup = BeautifulSoup(browser.page_source, "html.parser")

#Obtenemos las Urls de los 20 destinos mas populares de España
urls=[]
for i in soup.find_all( "div", class_="geo_name"):    
    print("https://www.tripadvisor.es" + i.find("a")["href"])
    urls.append( "https://www.tripadvisor.es"+ i.find("a")["href"]) 

#En esta etapa del proyecto solo se utilizaran los restaurantes de Madrid
print(urls[0])

# Con este codigo se manipula  la url de Madrid para que vaya a todas las paginas de restaurantes de Madrid
cadena= []
cadena = urls[0].split("-")
cadena

#salen 30 restaurantes por pagina    
url_rest_perpage=[]
for i in range(330):#330 es total de num indices que se observa en la pagina de restaurantes de madrid 
        nueva_cadena = cadena[0] + "-"+ cadena[1] + f"-oa{30*i}-"+ cadena[2]
        #print(nueva_cadena)
        url_rest_perpage.append(nueva_cadena)
print(url_rest_perpage[0:5])#ejemplo para que se vea

%%time
#Este codigo tarda 48 min!!!!!!!!!!!!!!
#Codigo para escoger los restaurantes por pagina
rest_url = []
for x, y in enumerate(url_rest_perpage):
    print(f"Procesando:{x+1} - URL: {y}")
    
    # Iniciar el navegador Helium de forma headless (sin mostrar la ventana del navegador)
    driver = helium.start_chrome(y, headless=True)    
    # Esperar unos segundos para que la página se cargue completamente
    sleep(2)    
    # Crear un objeto BeautifulSoup a partir del HTML de la página
    soup = BeautifulSoup(driver.page_source, "html.parser")    
    # Encontrar todos los elementos <div> con la clase "RfBGI"
    for i in soup.find_all("div", class_="RfBGI"):
        href = i.find("a")["href"]
        if href:
            rest_url.append(href)
            #print(f"{href}")
                    
    
    # Cerrar la ventana del navegador
    driver.quit()

# Imprimir la cantidad total de enlaces extraídos
print(f"Total de enlaces extraídos: {len(rest_url)}")

#para quitar los repetidos
restaurantes = list(set(rest_url))
len(restaurantes)


#Se arregla la cadena para convertirla en una URL
web_rest=[]
for i in range(len(restaurantes)):
    
    #Se guardan todas las URL de los restaurantes en la lista web_rest
    web_rest.append("https://www.tripadvisor.es" + f"{restaurantes[i]}")

len(web_rest)#este archivo ya a

#ejemplo de como queda
print(web_rest[15])

#La ejecucion del codigo anterior tarda 48 min, asi que guardamos el archivo en un pkl

# Especifica el nombre del archivo pickle y la carpeta de destino
nombre_archivo = "web_rest.pickle.pkl"
carpeta_destino = "Sources"

ruta_archivo = os.path.join(carpeta_destino, nombre_archivo)

# Guarda el archivo en modo de escritura binaria (wb)
with open(ruta_archivo, 'wb') as archivo:
    pickle.dump(web_rest, archivo)
    

#####################################################################
