==> ETL/2_Url_plantilla_WebScrapping_OnebyOne Etapa2.ipynb <==
Este Notebook es para obtener las URL de cada uno de los restaurantes de Madrid

import pandas as pd
import numpy as np
import helium
from time import sleep
from bs4 import BeautifulSoup
import pickle


#Empezamos trabajando la lista de direcciones url de restaurantessacadas del notebook 1_URL_ResT
#cuando se hizo el webscrapping este es el archivo original...ya ha comabiado
with open("Sources\web_rest.pickle", "rb") as f:
    rest_url = pickle.load(f)


len(rest_url)

#ejemplo de como queda
print(rest_url[50])

#Despues de varias pruebas el scrapping hay que hacerlo de 200 en 200
shorter_list=rest_url[9600:9800]
len(shorter_list)

%%time
Nombre=[]
Score=[]
Num_Reviews=[]
Posicion=[]
Rango_precio=[]
Tipo_Cocina=[]
Awards=[]
Score_Comida=[]
Score_Servicio=[]
Calidad_Precio=[]
Ambiente=[]
Direccion=[]
Horario=[]
Web_TripAdvisor=[]
Ultimo_review=[]
Opciones=[]
Excelente=[]
MuyBueno=[]
Normal=[]
Malo=[]
Pesimo=[]
Contacto=[]

c=0
for url in shorter_list:
    
    print(f"Faltan {len(shorter_list)- c } restaurantes")
    c+=1
    
    browser = helium.start_chrome(url, headless = True)
    soup = BeautifulSoup(browser.page_source, "html.parser")
    sleep(2)
    #Web
    Web_TripAdvisor.append(url)
    sleep(0.1)
    
    #Nombre##########
    try:
        nombre = soup.find("h1", class_="HjBfq").text
    except:
        nombre = np.nan
    Nombre.append(nombre)
    sleep(0.1)
    
    #Score#######################
    try:
        score= soup.find("span", class_="ZDEqb").text.replace("\xa0", "")
    except:
        score=np.nan
    Score.append(score)
    sleep(0.1)
    
    #Numero_Reviews####################
    try:
        num_review= soup.find("span", class_ = "AfQtZ").text.split()[0]
    except:
        num_review= np.nan
    Num_Reviews.append(num_review)
    sleep(0.1)
    
    #Posicion##################
    try:
        posicion= soup.find("div", class_="cNFlb").text.split()[1]
    except:
        posicion=np.nan
    Posicion.append(posicion)
    sleep(0.1)
    
    #Rango_precio##################
    try:
        rango_precio=soup.find_all("a", class_ = "dlMOJ")[0].text
    except:
        rango_precio=np.nan        
    Rango_precio.append(rango_precio)
    sleep(0.1)
    
    #Awards################
    try:
        awards=soup.find("span", class_="lgtpa").text
    except:
        awards="-"
    Awards.append(awards)
    sleep(0.1)
    
    #Tipo de cocina#########################
    try:
        tipo_cocina=[]
        for x in soup.find_all("a", class_="dlMOJ"):
            tipo_cocina.append(x.text)
        tipo_cocina=tipo_cocina[1:]
    except:
        tipo_cocina=np.nan
    Tipo_Cocina.append(tipo_cocina)
    sleep(0.1)
    
    #Score_comida#######
    try:
        punt_comida=soup.find_all('div', class_='DzMcu')[0]
        score_comida=float(punt_comida.find('span', 
                           class_='ui_bubble_rating').get('class')[-1].split('_')[-1])/10
    except:
        score_comida=np.nan
    Score_Comida.append(score_comida)
    sleep(0.1)
    
    #Score_servicio#######
    try:
        punt_servicio=soup.find_all('div', class_='DzMcu')[1]
        score_servicio=float(punt_servicio.find('span', 
                           class_='ui_bubble_rating').get('class')[-1].split('_')[-1])/10
    except:
        score_servicio=np.nan
    Score_Servicio.append(score_servicio)
    sleep(0.1)
    
    #Calidad_Precio#######
    try:
        punt_calidad=soup.find_all('div', class_='DzMcu')[2]
        calidad_precio=float(punt_calidad.find('span', 
                           class_='ui_bubble_rating').get('class')[-1].split('_')[-1])/10
    except:
        calidad_precio=np.nan
    Calidad_Precio.append(calidad_precio)
    sleep(0.1)
    
    #Ambiente#######
    try:
        punt_ambiente=soup.find_all('div', class_='DzMcu')[3]
        ambiente=float(punt_ambiente.find('span', 
                           class_='ui_bubble_rating').get('class')[-1].split('_')[-1])/10
    except:
        ambiente=np.nan
    Ambiente.append(ambiente)
    sleep(0.1)
    
    #Direccion#################
    try:
        direccion_detalles=soup.find_all("a", class_="AYHFM")
        direccion=direccion_detalles[1].text.split(",")
    except:
        direccion=np.nan
    Direccion.append(direccion)
    sleep(0.1)
    
    #Puntuaciones Viajero
    #Excelente
    try:
        excelente= soup.find_all("div", class_ = "ui_checkbox item")[0].text.split("Excelente")[-1]
    except:
        excelente=np.nan
    Excelente.append(excelente)
    sleep(0.1)
    
    #Muy bueno
    try:
        muy_bueno=soup.find_all("div", class_ = "ui_checkbox item")[1].text.split("Muy bueno")[-1]
    except:
        muy_bueno=np.nan
    MuyBueno.append(muy_bueno)
    sleep(0.1)
    
    #Bueno
    try:
        bueno=soup.find_all("div", class_ = "ui_checkbox item")[2].text.split("Normal")[-1]
    except:
        bueno=np.nan
    Normal.append(bueno)
    sleep(0.1)
    
    #Malo
    try:
        malo=soup.find_all("div", class_ = "ui_checkbox item")[3].text.split("Malo")[-1]
    except:
        malo=np.nan
    Malo.append(malo)
    sleep(0.1)
    
    #Pesimo
    try:
        pesimo=soup.find_all("div", class_ = "ui_checkbox item")[4].text.split("Pésimo")[-1]
    except:
        pesimo=np.nan
    Pesimo.append(pesimo)
    sleep(0.1)
          
    #Ultimo_review#########################
    try:
        comment=[]
        for i in soup.find_all("div", class_="entry"):
            comment.append(i.find_all("p")[0].text)
        #para quitar las "\n"
        for x in comment:
            cleaned_text = x.replace("\n", "")
        comentario=comment[0]
    except:
        comentario=np.nan
    Ultimo_review.append(comentario)
    sleep(0.1)  
    
    #Opciones##########################
    try:
        opc=[]
        for x in soup.find_all("div", class_ = "SrqKb"):
            opc.append(x.text.replace("\xa0", ""))
        opciones=opc[1:]
    except:
        opciones=np.nan
    Ultimo_review.append(opciones)
    sleep(0.1)    


df_TripAdvisor_Rest_Madrid=pd.DataFrame()

df_TripAdvisor_Rest_Madrid["Nombre"]=Nombre
df_TripAdvisor_Rest_Madrid["Score"]=Score
df_TripAdvisor_Rest_Madrid["Excelente"]=Excelente
df_TripAdvisor_Rest_Madrid["MuyBueno"]=MuyBueno
df_TripAdvisor_Rest_Madrid["Normal"]=Normal
df_TripAdvisor_Rest_Madrid["Malo"]=Malo
df_TripAdvisor_Rest_Madrid["Pesimo"]=Pesimo
df_TripAdvisor_Rest_Madrid["Num_Reviews"]=Num_Reviews
df_TripAdvisor_Rest_Madrid["Posicion"]=Posicion
df_TripAdvisor_Rest_Madrid["Rango_precio"]=Rango_precio
df_TripAdvisor_Rest_Madrid["Tipo_Cocina"]=Tipo_Cocina
df_TripAdvisor_Rest_Madrid["Awards"]=Awards
df_TripAdvisor_Rest_Madrid["Score_Comida"]=Score_Comida
df_TripAdvisor_Rest_Madrid["Score_Servicio"]=Score_Servicio
df_TripAdvisor_Rest_Madrid["Calidad_Precio"]=Calidad_Precio
df_TripAdvisor_Rest_Madrid["Ambiente"]=Ambiente
df_TripAdvisor_Rest_Madrid["Direccion"]=Direccion
df_TripAdvisor_Rest_Madrid["Web_TripAdvisor"]=Web_TripAdvisor


df_TripAdvisor_Rest_Madrid.head()

#Guardo el diccionario en un archivo pickle

# Especifica el nombre del archivo pickle y la carpeta de destino
nombre_archivo = "Trip_Rest_9800_9869.pkl"
carpeta_destino = "Sources/datos pkl"

ruta_archivo = os.path.join(carpeta_destino, nombre_archivo)

# Guarda el archivo en modo de escritura binaria (wb)
with open(ruta_archivo, 'wb') as archivo:
    # Utiliza la función dump de pickle para guardar el objeto en el archivo
    pickle.dump(df_TripAdvisor_Rest_Madrid, archivo)


################
