==> ETL/3_join_pkl_file.ipynb <==
import pickle
import pandas as pd
import numpy as np

def concat_pkl_files(file_paths):
    """
    Lee archivos pkl desde una lista de rutas de archivos,los convierte en DataFrames,
    y los concatena en uno solo.
    :param file_paths: Lista de rutas de archivos pkl.
    :return: DataFrame resultante de la concatenación.
    """
    dfs = []  # Lista para almacenar los DataFrames de cada archivo
    
    for file in file_paths:
        with open(f"Sources\datos pkl\{file}", "rb") as f:
            df = pickle.load(f)
        dfs.append(df)  # Agregar el DataFrame a la lista
    
    concatenated_df = pd.concat(dfs, ignore_index=True)  # Concatenar los DataFrames
    return concatenated_df

#Codigo para generer todos los nombres de archivo pkl
file_pkl=[]
for i in range(0,9800,200):
    file_pkl.append(f"Trip_Rest_{i}_{i+200}.pkl")
print(file_pkl)


result_df2 = concat_pkl_files(file_pkl)

result_df2.info()

with open("Sources\datos pkl\Trip_Rest_9800_9869.pkl", "rb") as f:
    df = pickle.load(f)

df_Trip_Rest_Madrid = pd.concat([result_df2, df], ignore_index=True)

df_Trip_Rest_Madrid.info()

df_Trip_Rest_Madrid.to_csv("Sources\Trip_Rest_Madrid.csv")
