==> ETL/4_Diccionario_clasificador_tipococina.ipynb <==
import pandas as pd
import numpy as np

import os
import pickle

from collections import Counter

# Abrimos la lista type_cocina

ruta_archivo = "Sources/type_cocina_list"

with open(ruta_archivo, "rb") as archivo:
    type_cocina_list = pickle.load(archivo)
    
print(type_cocina_list)


len(type_cocina_list)

#este es un Archivo que se creo en excel de guia para agrupar los tipos de cocina individuales
#a mas generales
df_clasificador=pd.read_excel("Sources/Clasificacion_ Tipos_ de_ Cocina.xlsx")

#Este df es solo para mostrar como se asigno la clasificacion
df_clasificador

#Creamos un diccionario para clasificarlas guiandonos por el csv anterior
dictionary = {}
for i in range(130):
    value = input(f"{type_cocina_list[i]} :Ingrese el valor correspondiente: ")
    dictionary[type_cocina_list[i]] = value

print("\nDiccionario resultante:")
for key, value in dictionary.items():
    print(f"{key}: {value}")

dictionary["Jamaicana"]="Caribeña"

#Guardo el diccionario en un archivo pickle
# Especifica el nombre del archivo pickle y la carpeta de destino
nombre_archivo = 'dic_resumen_tipo cocina.pickle'
carpeta_destino = 'Sources'

ruta_archivo = os.path.join(carpeta_destino, nombre_archivo)

# Guarda el archivo en modo de escritura binaria (wb)
with open(ruta_archivo, 'wb') as archivo:
    # Utiliza la función dump de pickle para guardar el objeto en el archivo
    pickle.dump(dictionary, archivo)
