==> ETL/4_Rango Precio.ipynb <==
import pandas as pd
import numpy as np
from time import sleep
import pickle
import os

### Rango de precio
para extraer de esta columna los tipos de cocina que aparecen

df=pd.read_csv("Sources\Trip_Rest_Madrid.csv")

df.info()

df["Rango_precio"].unique()

df_Precio=df[["Nombre", "Rango_precio", "Tipo_Cocina"]]

df_Precio.head(20)

df_Precio["Rango_precio"]=df_Precio["Rango_precio"].astype(str)


type(df_Precio["Rango_precio"][18])

#sustituir los valores de rango de precio que estan erroneos
#"€€ - €€€" se sustituye el valor por un precio medio
otros_tipococina={}
for i, j in enumerate(df_Precio["Rango_precio"]):
    if j!="€" and j!="€€ - €€€" and j!="€€€€":
            otros_tipococina[i]= j
            df_Precio["Rango_precio"][i]="€€ - €€€"#se sustituye el valor por un precio medio
 #Se almacena en un diccionario los tipos de cocina que por error estan en este df          
print(otros_tipococina)     
len(otros_tipococina)  

lista_rangoprecio_actualizado=list(df_Precio["Rango_precio"])


type(lista_rangoprecio_actualizado)

with open("Sources/lista_rangoprecio_actualizado", 'wb') as archivo:
    # Utiliza la función dump de pickle para guardar el objeto en el archivo
    pickle.dump(lista_rangoprecio_actualizado, archivo)

df_Precio.head(20)

#Guardo el diccionario en un archivo pickle

# Especifica el nombre del archivo pickle y la carpeta de destino
nombre_archivo = 'dicc_tipococina_rangoprecio.pkl'
carpeta_destino = 'Sources'

ruta_archivo = os.path.join(carpeta_destino, nombre_archivo)

# Guarda el archivo en modo de escritura binaria (wb)
with open(ruta_archivo, 'wb') as archivo:
    # Utiliza la función dump de pickle para guardar el objeto en el archivo
    pickle.dump(otros_tipococina, archivo)


#######################
