==> ETL/8_Recomendador_0.ipynb <==
import numpy as np
import pandas as pd
import math
import time

df=pd.read_csv("Sources/df_Clean_TripAdvisor.csv")

df.info()

#Hay que ordenar el df por las valoraciones para que ya se quede filtrado
#para eelo tenemos una funcion

#Hay que ajustar las columnas Score y Num_Reviews
""" Para calcular una métrica que capture la relación entre la puntuación y el 
número de comentarios (reviews) de un restaurante, puedes utilizar el concepto 
de "puntuación ponderada" donde le das más importancia al número de comentarios. 
Una forma simple de hacerlo es multiplicar la puntuación por una función del número 
de comentarios. Aquí tienes una posible fórmula:
Puntuación Ponderada = Puntuación * (1 + Factor * Log(Número de Comentarios + 1))
"""
#Ejemplo de como funcionaria la formula
factor=0.6 # se elige un factor para dar mayor o menor importancia a la puntuacion
Punt_poderada_list=[]
for x,y in zip(df["Score"][0:10], df["Num_Reviews"][0:10]):
    Punt_poderada= round(x *(1 + factor * math.log(y+ 1)),3)
    print(f"Score:{x},Num_Reviews:{y}, Punt_poderada:{Punt_poderada}")
    Punt_poderada_list.append(Punt_poderada)
    # explicar con indice 0 e indice 3 Teoria vs Punt_Ponderada

factor=0.6 # se elige un factor para dar mayor o menor importancia a la puntuacion
Punt_poderada_list=[]
for x,y in zip(df["Score"], df["Num_Reviews"]):
    Punt_poderada=round(x *(1 + factor * math.log(y+ 1)),3)   
    Punt_poderada_list.append(Punt_poderada)  

#con el resultado de la formula anterior se crea una nueva columna que se llama valoracion
df["Valoracion"]=Punt_poderada_list

df.info()

### Recomendador por Tipo Cocina

#codigo para saber si el usuario prefiere entrar a escoger el tipo de cocina
while True:
    try:
        recomienda_cocina = int(input("¿Tiene preferencias por algún tipo de cocina? (Si=1 / No=0): "))
        if recomienda_cocina != 0 and recomienda_cocina != 1:
            raise ValueError("Por favor, ingrese 0 o 1.")#evalua si alcanza otro valor diferente
        break  # Salir del bucle si la entrada es válida
    except ValueError:
        print("Entrada no válida. Por favor, ingrese 0 o 1.")


# Define las opciones que el usuario puede calificar
if recomienda_cocina:
    opciones = ["Latina", "Asiatica", "Italiana", "Española", 
            "America Norte", "Eur Este","Eur Occidental", "Arabe",
            "Africana", "India"]

# Crea un diccionario para almacenar las calificaciones
    calificaciones = {}    
# Solicita al usuario que califique cada opción 
    print("Por favor, califique qué tipo generico de cocina le gusta(escala 1 a 10),No se puede repetir la calificacion")
    print("Siendo 10: Me encanta y 1:No me apetece")
    for opcion in opciones:
        calificacion = int(input(f"Califique '{opcion}' (1-10): "))
        if calificacion > 0 and calificacion < 11:
            calificaciones[opcion] = calificacion/10
        
        else:
            print("La clasificacion es incorrecta")
            break
            
    print("\nCalificaciones:", calificaciones)

else:
    print("Ya que no tiene preferencias por tipo de cocina le mostramos un lista ordenada por valoracion")
    num_rest= int(input(f"¿Cuántos restaurantes quiere en su lista(1-50)"))
    df_ordenado = df.sort_values(by='Valoracion', ascending=False)
    df_sinpreferencia =df_ordenado[["Nombre", "Score","Valoracion","Tipo_Cocina"]].head(num_rest)#los 25
    print(df_sinpreferencia)

#Funcion ordena diccionarios
def ordena_diccionario(diccionario):
    # Ordenar el diccionario de mayor a menor
    diccionario_ordenado = dict(sorted(diccionario.items(), key=lambda item: item[1], reverse=True))
    
    return diccionario_ordenado   

 # Mostrar el diccionario ordenado
calificaciones_ordenado=ordena_diccionario(calificaciones)  
calificaciones_ordenado

#Funcion para puntuar los tipos de cocina segun deseo del usuario

def puntuacion_cocinas(lista):                  
    # Solicita al usuario que califique cada opción    
    print(f"Por favor, califique las opciones de cocina:{lista} (escala 1 a 10)")
    print("Siendo 10: Me encanta y 1:No me apetece")
    
    calificaciones1={}
    for opcion in lista:
                calificacion = int(input(f"Califique '{opcion}' (1-10): "))
                if calificacion > 0 and calificacion < 11:
                    calificaciones1[opcion] = calificacion/10
                else:
                    print("La clasificacion es incorrecta")
                    break 
    return calificaciones1

# Verifica si la clave deseada está entre las tres primeras posiciones
clave = ["Latina", "Asiatica"]  
primeras_tres_claves = list(calificaciones_ordenado.keys())[:3]
Latinoamerica =["America Sur","Asador", "Caribeña", "Latina", "Mexicana"]
Asiatica = ["Asiatica", "China", "Japonesa"]

for x in clave:
    if x in primeras_tres_claves:        
        if x == "Latina":
            print(f"\n Calificaciones {x}")
            calif_latina=puntuacion_cocinas(Latinoamerica)
            calif_latina_ordenado=ordena_diccionario(calif_latina)
            lista_latin= list(calif_latina_ordenado.keys())[0:2]
            #calif_latina_ordenado            
            #print(calif_latina)            
        elif x== "Asiatica":
            print(f"\n Calificaciones {x}")
            calif_asiatica=puntuacion_cocinas(Asiatica)
            calif_asiatica_ordenado=ordena_diccionario(calif_asiatica)
            lista_asia= list(calif_asiatica_ordenado.keys())[0:2]
            #calif_asiatica_ordenado
            #print(calif_asiatica)        
        
    else:
        lista_pref= list(calificaciones_ordenado.keys())[0:4]
        #print(f'La clave "{x}" no se encuentra entre las tres primeras posiciones.')

#Para organizar por orden de peso la lista y sublista de tipos de puntuacion
lista_pref= list(calificaciones_ordenado.keys())[0:4]
lista_pref
tipo_cocina_preferida_list=[]
for x in lista_pref:
    if x != "Latina" and x!="Asiatica":
        tipo_cocina_preferida_list.append(x)
    elif x == "Latina":
        tipo_cocina_preferida_list.extend(lista_latin)
        
    elif x == "Asiatica":
        tipo_cocina_preferida_list.extend(lista_asia)
print(tipo_cocina_preferida_list)
    

resultado=pd.DataFrame()
dicc_df={}
z= int(input(f"¿Cuantas recomendaciones quiere por tipo de cocina?(1-10)"))#numero de recomendaciones por tipo de cocina
for y in tipo_cocina_preferida_list:
    for x in df.columns[19:-4]:#se corresponde con las columnas tipo cocina
        if x == y:
            df_recomendador1 = df[df[x]==1]
            print(f"El tipo de cocina Seleccionada: {y}")
            df_ordenado = df_recomendador1.sort_values(by='Valoracion', ascending=False)
            print(df_ordenado[["Nombre", "Score","Valoracion"]].head(z))
            dicc_df[x]=df_ordenado[["Nombre", "Score","Valoracion"]].head(z)


# Funcion

z = tab1.number_input()

def recomendador_final(tipo_cocina_preferida_list, z):

    resultado=pd.DataFrame()
    dicc_df={}
#     z= int(input(f"¿Cuantas recomendaciones quiere por tipo de cocina?(1-10)"))#numero de recomendaciones por tipo de cocina
    for y in tipo_cocina_preferida_list:
        for x in df.columns[19:-4]:#se corresponde con las columnas tipo cocina
            if x == y:
                df_recomendador1 = df[df[x]==1]
                print(f"El tipo de cocina Seleccionada: {y}")
                df_ordenado = df_recomendador1.sort_values(by='Valoracion', ascending=False)
                print(df_ordenado[["Nombre", "Score","Valoracion"]].head(z))
                dicc_df[x]=df_ordenado[["Nombre", "Score","Valoracion"]].head(z)
                
    return dicc_df

list(dicc_df.values())[5]

#################################
